Fix compilation errors: 10 bugs squashed
- Fixed TaskMapper import (top-level extension functions)
- Removed redundant Long? type converters in Room
- Added missing styles Theme.KisloGTD.Splash and Theme.KisloGTD.Transparent
- Added missing Column and displayTitle imports across all screens
- Fixed PaddingValues constructor calls (ControlScreen, LaterScreen)
- Removed neutralButton from Material3 AlertDialog (DayScreen)
- Wrapped Flow.collect() in viewModelScope.launch (DashboardViewModel)
- Fixed MaterialTheme.DefaultColorScheme → colorScheme (CategoryIcon)
- Fixed JSONObject.getFloat → getDouble().toFloat() (BackupManager)
- Added Gradle wrapper files for build reproduction

+++ app/src/main/java/com/kozvits/kislogtd/sync/BackupManager.kt
@@ -143,7 +143,7 @@ internal fun weeklyStatsEntityFromJson(json: JSONObject): WeeklyStatsEntity {
         weekStartDate = json.getLong("weekStartDate"),
         totalCompleted = json.getInt("totalCompleted"),
         projectCompleted = json.getInt("projectCompleted"),
-        projectPercent = json.getFloat("projectPercent"),
+        projectPercent = json.getDouble("projectPercent").toFloat(),
         diaryEntry = json.optString("diaryEntry", "")
     )
 }
